Add files via upload
Movie review classification is the model which has been created to identify the sentiment from the given text. It is defined as a process in which the text mining techniques are extracted and further it is being studied ahead to understand the polarity of the given document or text. In this implemented principle, we are mainly focusing in understanding how accurate it is when we compare it with the human judgments and other conclusions to understand the similarity and focus. These understanding and comparisons are mainly calculated by different measures which is based on precision and recall oven three variables which are, Negative, Positive and Neutral.
While we consider from a business point of view, the final output from the prepared data can be helpful in many ways. it could be helpful in making a decision, it could be helpful in understanding end result of a political or a social media campaign etc. Here in this assignment, we are mainly focusing on implementing the sentimental analysis on movie sentiment classification